+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/network/serverpackets/AcquireSkillInfo.java
@@ -21,7 +21,7 @@ public class AcquireSkillInfo extends IClientOutgoingPacket {
     private final AcquireSkillType _type;
     private final int _id;
     private final int _level;
-    private final int _spCost;
+    private final long _spCost;
     private final List<Req> _reqs;
 
     /**
